--- a/eotsmanager/cmd/eotsd/main.go
+++ b/eotsmanager/cmd/eotsd/main.go
@@ -16,6 +16,6 @@ func main() {
 
 	if err := daemon.NewRootCmd().ExecuteContext(ctx); err != nil {
 		_, _ = fmt.Fprintf(os.Stderr, "Error while executing eotsd CLI: %s", err.Error())
-		os.Exit(1)
+		os.Exit(1) //nolint:gocritic
 	}
 }