This PR removes the usage of `signal.Interceptor` and uses Go's idiomatic `context` to handle cancellations. This enables us to run e2e tests in parallel (almost 4x faster).

```
//old-time
ok      github.com/babylonlabs-io/finality-provider/itest       548.976s
// new time
ok      github.com/babylonlabs-io/finality-provider/itest       139.579s
```

Time in GH action `9m 58s` vs `3m 24s`


Closes #219